+++ servers/libvirt/lib/Resources.php
@@ -8,21 +8,49 @@ class Resources
 
     public function __construct($serviceId) {
         $this->resource = Capsule::table('mod_libvirt_domains')
+            ->select(
+                'domain_id',
+                'mod_libvirt_domains.name',
+                'node_ip_address',                                
+                'tblservers.username',
+                'power_state',
+                'ram',                
+                'vcpus',                
+            )
             ->where('whmcs_service_id', $serviceId)
+            ->join('tblhosting', 'mod_libvirt_domains.whmcs_service_id', '=', 'tblhosting.id')
+            ->join('tblservers', 'tblhosting.server', '=', 'tblservers.id')            
             ->first();
-    }
-    
-    public function getVcpus()
-    {        
-        return $this->resource->vcpus;
+            // die(print_r($this->resource,1));
     }
 
-    public function getRam() {
+    public function domainId() {
+        return $this->resource->domain_id;
+    }
+    
+    public function name() {
+        return $this->resource->name;
+    }
+
+    public function nodeIpAddress() {
+        return $this->resource->node_ip_address;
+    }
+
+    public function nodeUsername() {
+        return $this->resource->username;
+    }
+                
+    public function powerState() {
+        return ucfirst($this->resource->power_state);
+    }
+
+    public function ram() {
         return $this->resource->ram;
     }
 
-    public function getPowerState() {
-        return ucfirst($this->resource->power_state);
+    public function vcpus()
+    {        
+        return $this->resource->vcpus;
     }
 
 }